requirements.txt
gunicorn==20.0.4
django==3.1
django-model-utils==4.0.0
django-environ
psycopg2-binary==2.9.2
django-debug-toolbar==2.2
